organizations/org-01/workers/worker_b.py
import React from 'react';

const WorkerB: React.FC = () => {
  return (
    <div className="bg-white dark:bg-gray-800 p-4">
      <Card className="w-full bg-card">
        <CardHeader className="bg-card">
          <CardTitle className="bg-card">Worker B AIエージェント</CardTitle>
          <CardDescription className="bg-card">
            バックエンド開発専門のWorkerエージェント
          </CardDescription>
        </CardHeader>
        <CardContent className="bg-card">
          <ul className="list-disc pl-5">
            <li>BaseAgentクラスからの継承とPython実装</li>
            <li>API開発・データベース設計・サーバー構築対応</li>
            <li>Python/Java/Go等のバックエンド言語対応</li>
            <li>Bossからの指示受信と実行結果の報告</li>
            <li>セキュリティ・パフォーマンス最適化</li>
            <li>インフラ構築とDevOps対応</li>
            <li>障害対応と復旧処理</li>
          </ul>
        </CardContent>
      </Card>
    </div>
  );
};

export default WorkerB;