pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "haconiwa"
version = "0.1.0"
description = "AI協調開発支援Python CLIツール - 箱庭での開発を支援"
readme = "README.md"
authors = [
    {name = "Daisuke Motoki", email = "kanri@kandaquantum.co.jp"}
]
requires-python = ">=3.8"
license = {text = "MIT"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "typer>=0.9.0",
    "libtmux>=0.23.2",
    "gitpython>=3.1.40",
    "prometheus-client>=0.17.1",
    "pyyaml>=6.0.1",
    "rich>=13.7.0",
    "click>=8.1.7",
]

[project.optional-dependencies]
docker = ["docker>=6.1.3"]
dev = [
    "pytest>=7.4.3",
    "pytest-cov>=4.1.0",
    "black>=23.11.0",
    "flake8>=6.1.0",
    "mypy>=1.7.1",
    "bandit>=1.7.5",
    "isort>=5.12.0",
    "pre-commit>=3.5.0",
]

[project.urls]
Homepage = "https://github.com/haconiwa/haconiwa"
Documentation = "https://haconiwa.readthedocs.io"
Repository = "https://github.com/haconiwa/haconiwa.git"
Issues = "https://github.com/haconiwa/haconiwa/issues"

[project.scripts]
haconiwa = "haconiwa.cli:app"

[tool.black]
line-length = 88
target-version = ["py38"]

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.8"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --cov=haconiwa"
testpaths = ["tests"]